pyproject.toml
[tool.poetry]
name = "static-website"
version = "0.1.0"
description = ""
authors = ["Joachim Jablon <ewjoachim@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
combine = "^2.5.1"
python-frontmatter = "^1.0.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
